support norm dyn shpae

+++ mindspore/python/mindspore/ops/function/array_func.py
@@ -5680,9 +5680,9 @@ def max(input, axis=None, keepdims=False, *, initial=None, where=None):    # pyl
         [0.7] [3]
     """
     if not input.shape:
-        return (input, Tensor(0))
+        return (input, Tensor(0, dtype=input.dtype))
     if axis is None:
-        return (reduce_max(input), Tensor(0))
+        return (reduce_max(input), Tensor(0, dtype=input.dtype))
     if initial is not None and not isinstance(initial, numbers.Number):
         raise TypeError(f"For 'max', 'initial' must be a scalar, but got {type(initial)}")
     if axis is not None and not isinstance(axis, int):
@@ -5786,9 +5786,9 @@ def min(input, axis=None, keepdims=False, *, initial=None, where=None):    # pyl
         [0.0] [0]
     """
     if not input.shape:
-        return (input, Tensor(0))
+        return (input, Tensor(0, dtype=input.dtype))
     if axis is None:
-        return (reduce_min(input), Tensor(0))
+        return (reduce_min(input), Tensor(0, dtype=input.dtype))
     if initial is not None and not isinstance(initial, numbers.Number):
         raise TypeError(f"For 'min', 'initial' must be a scalar, but got {type(initial)}")
     if axis is not None and not isinstance(axis, int):
